refactor : 기존의 로고를 모두 삭제할 경우에 대한 로직 추가

+++ src/main/java/com/epris/homepage/eprian/corporate_logo/service/LogoService.java
@@ -45,7 +45,7 @@ public class LogoService {
             }
         }
         /* 기존 로고 삭제 */
-        deleteLogo(deleteLogoList);
+        if(!deleteLogoList.isEmpty())deleteLogo(deleteLogoList);
 
         return findLogoListByType(type);
     }
